backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pypdf
python-dotenv
youtube-transcript-api

langchain
langgraph
langchain-core
langchain-groq

groq
pillow



